Move all models to util/models.py
