+++ src/murfey/server/api.py
@@ -4,21 +4,32 @@ import datetime
 import logging
 from functools import lru_cache
 from pathlib import Path
-from typing import Any, Dict, List, Optional
+from typing import List
 
 import packaging.version
 from fastapi import APIRouter, Request
 from fastapi.responses import HTMLResponse
 from ispyb.sqlalchemy import BLSession, Proposal
-from pydantic import BaseModel, BaseSettings
+from pydantic import BaseSettings
 
 import murfey.server
 import murfey.server.bootstrap
 import murfey.server.ispyb
 import murfey.server.websocket as ws
-import murfey.util.models
 from murfey.server import _transport_object, get_hostname, get_microscope, templates
 from murfey.server.config import from_file
+from murfey.util.models import (
+    ContextInfo,
+    DCGroupParameters,
+    DCParameters,
+    File,
+    ProcessFile,
+    ProcessingJobParameters,
+    RegistrationMessage,
+    SuggestedPathParameters,
+    TiltSeries,
+    Visit,
+)
 
 log = logging.getLogger("murfey.server.api")
 
@@ -90,13 +101,13 @@ def all_visit_info(request: Request, db=murfey.server.ispyb.DB):
         )
 
 
-@router.get("/demo/visits_raw", response_model=List[murfey.util.models.Visit])
+@router.get("/demo/visits_raw", response_model=List[Visit])
 def get_current_visits_demo(db=murfey.server.ispyb.DB):
     microscope = "m12"
     return murfey.server.ispyb.get_all_ongoing_visits(microscope, db)
 
 
-@router.get("/visits_raw", response_model=List[murfey.util.models.Visit])
+@router.get("/visits_raw", response_model=List[Visit])
 def get_current_visits(db=murfey.server.ispyb.DB):
     microscope = get_microscope()
     return murfey.server.ispyb.get_all_ongoing_visits(microscope, db)
@@ -146,11 +157,6 @@ def visit_info(request: Request, visit_name: str, db=murfey.server.ispyb.DB):
         return None
 
 
-class ContextInfo(BaseModel):
-    experiment_type: str
-    acquisition_software: str
-
-
 @router.post("/visits/{visit_name}/context")
 async def register_context(context_info: ContextInfo):
     log.info(
@@ -163,13 +169,6 @@ async def register_context(context_info: ContextInfo):
     )
 
 
-class File(BaseModel):
-    name: str
-    description: str
-    size: int
-    timestamp: float
-
-
 @router.post("/visits/{visit_name}/files")
 async def add_file(file: File):
     message = f"File {file} transferred"
@@ -178,31 +177,12 @@ async def add_file(file: File):
     return file
 
 
-class RegistrationMessage(BaseModel):
-    registration: str
-    params: Optional[Dict[str, Any]] = None
-
-
 @router.post("/feedback")
 async def send_murfey_message(msg: RegistrationMessage):
     if _transport_object:
         _transport_object.transport.send(
             "murfey_feedback", {"register": msg.registration}
         )
-
-
-class ProcessFile(BaseModel):
-    path: str
-    description: str
-    size: int
-    timestamp: float
-    processing_job: int
-    data_collection_id: int
-    image_number: int
-    mc_uuid: int
-    movie_uuid: int
-    autoproc_program_id: int
-    pixel_size: float
 
 
 @router.post("/visits/{visit_name}/tomography_preprocess")
@@ -252,16 +232,6 @@ async def request_tomography_preprocessing(visit_name: str, proc_file: ProcessFi
         return proc_file
     await ws.manager.broadcast(f"Pre-processing requested for {ppath.name}")
     return proc_file
-
-
-class TiltSeries(BaseModel):
-    name: str
-    file_tilt_list: str
-    dcid: int
-    processing_job: int
-    autoproc_program_id: int
-    motion_corrected_path: str
-    movie_id: int
 
 
 @router.post("/visits/{visit_name}/align")
@@ -325,10 +295,6 @@ def shutdown():
     return {"success": True}
 
 
-class SuggestedPathParameters(BaseModel):
-    base_path: Path
-
-
 @router.post("/visits/{visit_name}/suggested_path")
 def suggest_path(visit_name, params: SuggestedPathParameters):
     count: int | None = None
@@ -338,28 +304,6 @@ def suggest_path(visit_name, params: SuggestedPathParameters):
         count = count + 1 if count else 2
         check_path = check_path.parent / f"{check_path_name}{count}"
     return {"suggested_path": check_path}
-
-
-class DCGroupParameters(BaseModel):
-    experiment_type: str
-
-
-class DCParameters(BaseModel):
-    voltage: float
-    pixel_size_on_image: str
-    experiment_type: str
-    image_size_x: int
-    image_size_y: int
-    tilt: int
-    file_extension: str
-    acquisition_software: str
-    image_directory: str
-    tag: str
-
-
-class ProcessingJobParameters(BaseModel):
-    tag: str
-    recipe: str
 
 
 @router.post("/visits/{visit_name}/register_data_collection_group")
